learn_openai_agent_base_with_feishu_bot/requirements.txt
lark-oapi>=1.4.8
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.12.0
pydantic>=2.0.0
openai-agents>=0.1.0